2D RPG that inspired by the nostalgic game called the Legend of the Zelda
Peter: Implemented some puzzles in the game
Peter: Implemented Inventory System that is fully functional with swap, pick up and stack items features.
Peter: Finished NPC Dialogue System that helps the player interact with NPCs
Peter: Finished Puzzle 3 where player has to step on the correct path to solve the puzzle with sound-track from Bleach.
Christopher: Created the overall interaction with the enviroment and how there are some objects to be interacted with
Christopher: Puzzle on how they sprite is able to interact in a certain a pattern and have music implemented
Another member: Started on the shoping UI with currency system